+++ preload.js
@@ -1,7 +1,8 @@
 const { contextBridge, ipcRenderer } = require('electron');
 
 contextBridge.exposeInMainWorld('electronAPI', {
-  toggleFullscreen: () => ipcRenderer.send('toggle-fullscreen')
+  toggleFullscreen: () => ipcRenderer.send('toggle-fullscreen'),
+  openNmrPreview: () => ipcRenderer.send('open-nmr-preview')
 });
 
 window.addEventListener('DOMContentLoaded', () => {